rewrite the multi-step into a more robust form
This should be able to handle multiple sessions, including if a session was ended pre-maturely.

I'll go ahead and make one more change to `select_verse_range` to display the current page of the Bible that I am on.
Then I will make handling for pagination, which is to be done after the last verse of Revelation is done.